% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnomalyDetection-package.R
\docType{package}
\name{AnomalyDetection}
\alias{AnomalyDetection}
\alias{AnomalyDetection-package}
\title{Anomaly Detection Using Seasonal Hybrid Extreme Studentized Deviate Test}
\description{
A technique for detecting anomalies in seasonal univariate time series.
The methods uses are robust, from a statistical standpoint, in the presence of
seasonality and an underlying trend. These methods can be used in
wide variety of contexts. For example, detecting anomalies in system metrics after
a new software release, user engagement post an 'A/B' test, or for problems in
econometrics, financial engineering, political and social sciences.
}
\author{
Owen S. Vallis, Jordan Hochenbaum, Arun Kejariwal; Modernization
        contributions by Bob Rudis
}
