% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_anom_detection.R
\name{AnomalyDetectionTs}
\alias{AnomalyDetectionTs}
\alias{ad_ts}
\title{Anomaly Detection Using Seasonal Hybrid ESD Test}
\usage{
AnomalyDetectionTs(x, max_anoms = 0.1, direction = "pos", alpha = 0.05,
  only_last = NULL, threshold = "None", e_value = FALSE,
  longterm = FALSE, piecewise_median_period_weeks = 2, verbose = FALSE,
  na.rm = FALSE)

ad_ts(x, max_anoms = 0.1, direction = "pos", alpha = 0.05,
  only_last = NULL, threshold = "None", e_value = FALSE,
  longterm = FALSE, piecewise_median_period_weeks = 2, verbose = FALSE,
  na.rm = FALSE)
}
\arguments{
\item{x}{Time series as a two column data frame where the first column consists of the
timestamps and the second column consists of the observations.}

\item{max_anoms}{Maximum number of anomalies that S-H-ESD will detect as a percentage of the
data.}

\item{direction}{Directionality of the anomalies to be detected. One of:
\code{pos}, \code{neg}, \code{both}.}

\item{alpha}{The level of statistical significance with which to accept or reject anomalies.}

\item{only_last}{Find and report anomalies only within the last day or hr in the time seriess.
One of \code{NULL}, \code{day}, \code{hr}.}

\item{threshold}{Only report positive going anoms above the threshold specified. One of:
\code{None}, \code{med_max}, \code{p95}, \code{p99}.}

\item{e_value}{Add an additional column to the anoms output containing the expected value.}

\item{longterm}{Increase anom detection efficacy for time series that are greater than a month.
See `Details`` below.}

\item{piecewise_median_period_weeks}{The piecewise median time window as described in Vallis,
Hochenbaum, and Kejariwal (2014). Defaults to 2.}

\item{verbose}{Enable debug messages}

\item{na.rm}{Remove any NAs in timestamps.(default: \code{FALSE})}
}
\value{
The returned value is a data frame containing timestamps, values,
and optionally expected values.
}
\description{
A technique for detecting anomalies in seasonal univariate time series where the input is a
series of <timestamp, count> pairs.
}
\examples{
data(raw_data)

ad_ts(raw_data, max_anoms=0.02, direction='both')

# To detect only the anomalies on the last day, run the following:

ad_ts(raw_data, max_anoms=0.02, direction='both', only_last="day")
}
\references{
\itemize{
\item Vallis, O., Hochenbaum, J. and Kejariwal, A., (2014)
"A Novel Technique for Long-Term Anomaly Detection in the Cloud", 6th USENIX, Philadelphia, PA.
(\url{https://www.usenix.org/system/files/conference/hotcloud14/hotcloud14-vallis.pdf})
\item Rosner, B., (May 1983), "Percentage Points for a Generalized ESD Many-Outlier
Procedure", Technometrics, 25(2), pp. 165-172. (\url{https://www.jstor.org/stable/1268549})
}
}
\seealso{
\code{\link[=ad_vec]{ad_vec()}}
}
