% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_anom_detection.R
\name{AnomalyDetectionVec}
\alias{AnomalyDetectionVec}
\alias{ad_vec}
\title{Anomaly Detection Using Seasonal Hybrid ESD Test}
\usage{
AnomalyDetectionVec(x, max_anoms = 0.1, direction = "pos", alpha = 0.05,
  period = NULL, only_last = FALSE, threshold = "None", e_value = FALSE,
  longterm_period = NULL, verbose = FALSE)

ad_vec(x, max_anoms = 0.1, direction = "pos", alpha = 0.05,
  period = NULL, only_last = FALSE, threshold = "None", e_value = FALSE,
  longterm_period = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Time series as a column data frame, list, or vector, where the column consists of 
the observations.}

\item{max_anoms}{Maximum number of anomalies that S-H-ESD will detect as a percentage of the
data.}

\item{direction}{Directionality of the anomalies to be detected. One of:
\code{pos}, \code{neg}, \code{both}.}

\item{alpha}{The level of statistical significance with which to accept or reject anomalies.}

\item{period}{Defines the number of observations in a single period, and used during seasonal 
decomposition.}

\item{only_last}{Find and report anomalies only within the last period in the time series.}

\item{threshold}{Only report positive going anoms above the threshold specified. One of:
\code{None}, \code{med_max}, \code{p95}, \code{p99}.}

\item{e_value}{Add an additional column to the anoms output containing the expected value.}

\item{longterm_period}{Defines the number of observations for which the trend can be considered 
flat. The value should be an integer multiple of the number of observations in a single period. 
This increases anom detection efficacy for time series that are greater than a month.}

\item{verbose}{Enable debug messages}
}
\value{
The returned value is a list with the following components.

Data frame containing index, values, and optionally expected values.
}
\description{
A technique for detecting anomalies in seasonal univariate time series where the input is a 
series of observations.
}
\examples{
data(raw_data)

ad_vec(raw_data[,2], max_anoms=0.02, period=1440, direction='both')

# To detect only the anomalies in the last period, run the following:

ad_vec(
  raw_data[,2], max_anoms=0.02, period=1440, direction='both', only_last=TRUE
)
}
\references{
\itemize{
\item Vallis, O., Hochenbaum, J. and Kejariwal, A., (2014)
"A Novel Technique for Long-Term Anomaly Detection in the Cloud", 6th USENIX, Philadelphia, PA.
(\url{https://www.usenix.org/system/files/conference/hotcloud14/hotcloud14-vallis.pdf})
\item Rosner, B., (May 1983), "Percentage Points for a Generalized ESD Many-Outlier
Procedure", Technometrics, 25(2), pp. 165-172. (\url{https://www.jstor.org/stable/1268549})
}
}
\seealso{
\code{\link[=ad_ts]{ad_ts()}}
}
