% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{ase}
\alias{ase}
\title{Function to perform graph adjacency spectral embedding (ASE)}
\usage{
ase(
  A,
  d = NULL,
  d.max = round(log(nrow(A))),
  diagaug = TRUE,
  approx = TRUE,
  elbow = 2
)
}
\arguments{
\item{A}{adjacency matrix}

\item{d}{number of embedding dimension. If NULL, dimension is chosen automatically.}

\item{d.max}{maximum number of embedding dimensions to try when d is not provided. Default is round(log(nrow(A))).}

\item{diagaug}{whether to do diagonal augmentation (TRUE/FALSE)}

\item{approx}{whether to find a few approximate singular values and corresponding singular vectors of a matrix using irlba package (TRUE/FALSE).}

\item{elbow}{number of elbow selected in Zhu & Ghodsi method for the scree plot of each individual graph singular values. Default is \eqn{2}.}
}
\value{
A matrix with n rows and d columns containing the estimated latent positions
}
\description{
Function to perform graph adjacency spectral embedding (ASE)
}
\references{
Zhu, Mu and Ghodsi, Ali (2006), Automatic dimensionality selection from
the scree plot via the use of profile likelihood, Computational
Statistics & Data Analysis, Volume 51 Issue 2, pp 918-930, November, 2006.

Sussman, D.L., Tang, M., Fishkind, D.E., Priebe, C.E.  A
Consistent Adjacency Spectral Embedding for Stochastic Blockmodel Graphs,
\emph{Journal of the American Statistical Association}, Vol. 107(499), 2012
}
