% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{buildOmni}
\alias{buildOmni}
\title{Function to build OMNI matrix}
\usage{
buildOmni(Alist, diagaug = FALSE)
}
\arguments{
\item{Alist}{a list (length M) of n x n adjacency matrices or igraph objects}

\item{diagaug}{whether to do diagonal augmentation (TRUE/FALSE)}
}
\value{
Omnibus matrix of Mn x Mn
}
\description{
Function to build OMNI matrix
}
