% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{diagAug}
\alias{diagAug}
\title{Function to perform diagonal augmentation for graph adjacency matrix.}
\usage{
diagAug(A)
}
\arguments{
\item{A}{a hollow adjacency matrix}
}
\value{
a non-hollow adjacency matrix
}
\description{
Function to perform diagonal augmentation for graph adjacency matrix.
}
