% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{doMase}
\alias{doMase}
\title{Function to calculate \eqn{l_2} distance between adjacent latent position estmate for graphs and vertices with MASE}
\usage{
doMase(glist, latpos.list, nmase = 2, dsvd = NULL, attrweight = NULL)
}
\arguments{
\item{glist}{a list (length M) of n x n adjacency matrices or igraph objects}

\item{latpos.list}{a list (length M) of ASE estimates for each graphs}

\item{nmase}{number of graphs to do joint embedding of MASE. It can only be 2 or M.}

\item{dsvd}{dsvd is number of dimension to do joint svd for MASE. If NULL then dimension is chosen automatically as the second elbow selected in Zhu & Ghodsi method for the scree plot of the singular values of the concatenated spectral embeddings of invidual ASE estimates.}
}
\value{
A list containing a vector tnorm of length M-1, with the
latent position estmate difference for graphs, and a matrix pdist
 with latent position estmate difference for vertices (size n x M-1 ).
}
\description{
Function to calculate \eqn{l_2} distance between adjacent latent position estmate for graphs and vertices with MASE
}
