% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{fast2buildOmni}
\alias{fast2buildOmni}
\title{Function to build OMNI matrix with two matrix}
\usage{
fast2buildOmni(Alist, diagaug = FALSE, attrweight = NULL)
}
\arguments{
\item{Alist}{a list (length \eqn{2}) of n x n adjacency matrices or igraph objects}

\item{diagaug}{whether to do diagonal augmentation (TRUE/FALSE)}
}
\value{
Omnibus matrix of 2n x 2n
}
\description{
Function to build OMNI matrix with two matrix
}
