% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{fast2doOmni}
\alias{fast2doOmni}
\title{Function to calculate \eqn{l_2} distance between adjacent latent position estmate for graphs and vertices with OMNI}
\usage{
fast2doOmni(n, Z)
}
\arguments{
\item{n}{number of vertices in each graph}

\item{Z}{a matrix of size 2n x d as the latent estimate for the omnibus matrix contructed by adjacent graphs}
}
\value{
A list containing a numeric value tnorm, with the \eqn{l_2} norm of
latent position estmate difference for adjacent graphs, and a vector pdist
 with \eqn{l_2} distance between latent position estmates for each vertex (size n).
}
\description{
Function to calculate \eqn{l_2} distance between adjacent latent position estmate for graphs and vertices with OMNI
}
