% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate.tsg}
\alias{generate.tsg}
\title{Sample a time series of RDPG graph (length tmax > 17) with same 1-1 matched vertices unweighted
hollow symmetric undirected graphs, the latent positions i.i.d uniform.
Some vertices in 16-th and 17-th graphs are given perturbations so there exists anomalies at 16:17.}
\usage{
generate.tsg(n, nperturb, cperturb = NULL, rmin, rmax, tmax)
}
\arguments{
\item{n}{number of vertices}

\item{nperturb}{number of perturbed vertices}

\item{cperturb}{number of perturbation. Larger cperturb means more obvious anomalies.}

\item{rmin, rmax}{parameter for uniform[rmin, rmax].}

\item{tmax}{number of graphs must be greater than 17.}
}
\value{
A list containing a vector tnorm of length tmax-1, with the
latent position difference for graphs, and a matrix pdist
 with latent position estmate difference for vertices (size n x tmax-1 ),
 and a list of length tmax of undirectected hollow symmeric unweighted graphs
}
\description{
Sample a time series of RDPG graph (length tmax > 17) with same 1-1 matched vertices unweighted
hollow symmetric undirected graphs, the latent positions i.i.d uniform.
Some vertices in 16-th and 17-th graphs are given perturbations so there exists anomalies at 16:17.
}
\examples{

# Sample a time series of RDPG graph (length tmax > 17) with same 1-1 matched vertices unweighted
# hollow symmetric undirected graphs, the latent positions i.i.d uniform.
# Some vertices in 16-th and 17-th graphs are given perturbations so there exists anomalies at 16:17.
n <- 100 #number of vertices
nperturb <- 20 #number of perturbed vertices
cperturb <- .12 #number of perturbation, larger cperturb means more obvious anomalies.
rmin <- .2 # parameter for uniform[rmin, rmax].
rmax <- .8 # parameter for uniform[rmin, rmax].
tmax <- 22 # number of graphs must be greater than 17.
#Generate data or load the data you want
glist <- generate.tsg(n, nperturb, cperturb=NULL, rmin, rmax, tmax)$glist
}
