% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_factors.R
\name{get_all_factors}
\alias{get_all_factors}
\title{Find All Factors}
\source{
\url{http://stackoverflow.com/a/6425597/3851274}
}
\usage{
get_all_factors(n)
}
\arguments{
\item{n}{number to be factored}
}
\value{
A list containing the integer vector(s) containing all factors for the given
\code{n} inputs.
}
\description{
\code{get_all_factors} finds all factor pairs for a given integer (i.e. a number
that divides evenly into another number).
}
\examples{

# Find all the factors of 39304
get_all_factors(39304)

}
