% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getdegchange}
\alias{getdegchange}
\title{It extracts (non-zero) degree change \eqn{deg.change} matrix n \times m-1 from a list of graphs.}
\usage{
getdegchange(gip)
}
\arguments{
\item{gip}{a list of graphs in \code{igraph} format.}
}
\value{
A matrix of size n x t-1, with each element to be degree changes
}
\description{
It extracts (non-zero) degree change \eqn{deg.change} matrix n \times m-1 from a list of graphs.
}
\author{
Guodong Chen <gchen35@jhu.edu>
}
