% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getweightchange}
\alias{getweightchange}
\title{get weighted degree change for a list of weighted graphs
It extracts (non-zero) weighted degree change matrix n by m-1 \eqn{deg.change} from a list of graphs.}
\usage{
getweightchange(gip)
}
\arguments{
\item{gip}{a list of graphs in \code{igraph} format.}
}
\value{
A matrix of size n x t-1, with each element to be weight degree changes
}
\description{
get weighted degree change for a list of weighted graphs
It extracts (non-zero) weighted degree change matrix n by m-1 \eqn{deg.change} from a list of graphs.
}
