% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{giant.component}
\alias{giant.component}
\title{find largest connected component in a graph
It extracts (non-zero) largest connected subgraph .}
\usage{
giant.component(graph, ...)
}
\arguments{
\item{graph}{a graph in \code{igraph} format}
}
\description{
find largest connected component in a graph
It extracts (non-zero) largest connected subgraph .
}
