% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horns_curve.R
\name{horns_curve}
\alias{horns_curve}
\title{Horn's Parallel Analysis}
\usage{
horns_curve(N, p)
}
\arguments{
\item{N}{integer value}

\item{p}{integer value}
}
\value{
A vector of length \code{p} with the computed average eigenvalues
}
\description{
\code{horns_curve} computes the average eigenvalues produced by a Monte Carlo
simulation that randomly generates a large number of matrices of size \code{N} x \code{p},
where each element is drawn from a standard normal probability distribution.
The values can then be plotted or compared to the true eigenvalues of a dataset
for a dimensionality assessment.
}
\examples{

\dontrun{
# Perform Horn's Parallel analysis with matrix \\code{x} dimensions
x <- matrix(rnorm(200*3), ncol = 10)
N <- nrow(x)
p <- ncol(x)
horns_curve(N, p)

# Graph the scree line for a dimensionality assessment
curvepoints <- horns_curve(N, p)
plot(curvepoints)
}

}
\references{
J. L. Horn, "A rationale and test for the number of factors in factor analysis," Psychometrika, vol. 30, no. 2, pp. 179-185, 1965.
}

