% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_inspect.R
\name{inspect_block}
\alias{inspect_block}
\title{Block Inspection}
\usage{
inspect_block(data, block_length)
}
\arguments{
\item{data}{data}

\item{block_length}{integer value to divide data}
}
\value{
A list where each item is a data frame that contains the original data
  for each block denoted in the state vector.
}
\description{
\code{inspect_block} creates a list where the original data has been divided
  into blocks denoted in the state vector. Streamlines the process of inspecting
  specific blocks of interest.
}
\examples{

inspect_block(security_logs, 30)

}
\seealso{
\code{\link{tabulate_state_vector}} for creating the state vector matrix based
  on desired blocks.
}
