% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{jlcc}
\alias{jlcc}
\title{remove edges which has zero weights for all graphs (if any) and find jointly largest connected component in graphs. Finally it removes all self-loops,
It extracts (non-zero) igraph list \eqn{gip} and removes all edges with zero edge weights and return a list of jointly largest connected component in graphs without self-loops .}
\usage{
jlcc(gip)
}
\arguments{
\item{gip}{a list of graphs in \code{igraph} format}
}
\value{
A list of graphs in igraph format
}
\description{
remove edges which has zero weights for all graphs (if any) and find jointly largest connected component in graphs. Finally it removes all self-loops,
It extracts (non-zero) igraph list \eqn{gip} and removes all edges with zero edge weights and return a list of jointly largest connected component in graphs without self-loops .
}
