% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{mase.latent}
\alias{mase.latent}
\title{Function to obtain latent position estmates for MASE (Multiple Adjancency Spectral Embedding) with individual ASE estimates for corresponding graphs}
\usage{
mase.latent(A, latpos.list, dsvd = NULL)
}
\arguments{
\item{A}{a list (length M-1) of n x n adjacency matrices}

\item{latpos.list}{a list (length M-1) of individual ASE estimates for the corresponding adjacency matrices}

\item{dsvd}{dimension for joint embedding. If NULL then dimension is chosen automatically as the second elbow selected in Zhu & Ghodsi method for the scree plot of the singular values of the concatenated spectral embeddings of invidual ASE estimates.}
}
\value{
A list with length (M-1) consists of matrices with size n x dsvd each.
}
\description{
Function to obtain latent position estmates for MASE (Multiple Adjancency Spectral Embedding) with individual ASE estimates for corresponding graphs
}
