% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{mase}
\alias{mase}
\title{Function to perform joint embedding part of MASE}
\usage{
mase(
  Adj_list,
  latpos.list,
  dsvd = NULL,
  elbow_mase = 2,
  show.scree.results = FALSE
)
}
\arguments{
\item{Adj_list}{a list of adjacency matrices with the same size n x n}

\item{latpos.list}{Individual ASE estimate for the adjacent graphs}

\item{elbow_mase}{number of elbow selected in Zhu & Ghodsi method for the scree plot of the singular values of the concatenated spectral embeddings of MASE.}

\item{show.scree.results}{when TRUE, the histogram of the estimated d for each graph, and the scree plot of the singular values of  the graph is shown if d is not specified.}

\item{d}{number of joint embedding dimensions. If NA, dimension is chosen automatically}
}
\value{
A list containing a matrix V of size n x d, with the
estimated invariant subspace, and a list R with the individual score parameters for each graph (size d x d each).
}
\description{
Function to perform joint embedding part of MASE
}
\references{

}
