% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_adjust.R
\name{mc_adjust}
\alias{mc_adjust}
\title{Multi-Collinearity Adjustment}
\usage{
mc_adjust(data, delta1 = 0.1, delta2 = 0.9, action = exclude)
}
\arguments{
\item{data}{numeric data}

\item{delta1}{numeric value between 0-1 for the column variance (default = 0.1)}

\item{delta2}{numeric value between 0-1 for the correlation values (default = 0.9)}

\item{action}{select action for handling columns causing multi-collinearity issues
\enumerate{
   \item \code{exclude}: exclude all columns causing multi-collinearity issues (default)
   \item \code{select}: identify the columns causing multi-collinearity issues
     and allow the user to interactively select those columns to remove
 }}
}
\value{

}
\description{
\code{mc_adjust} handles issues with multi-collinearity.
}
\details{
\code{mc_adjust} handles issues with multi-collinearity by first removing
any columns whose variance is close to \eqn{0 + Δ1}. Then, it removes linearly
dependent columns. Finally, it removes any columns that have a high absolute
correlation value between \eqn{1 - Δ2} and \eqn{1}.
}
\examples{


}
