% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{pdistXY}
\alias{pdistXY}
\title{Calculate \eqn{l_2} distance between latent positions for vertices.}
\usage{
pdistXY(X, Y)
}
\arguments{
\item{X, Y}{are matrices with n rows and d columns}
}
\value{
A vector of size \eqn{n}, with element being \eqn{l_2} distance between rows of X and Y.
}
\description{
Calculate \eqn{l_2} distance between latent positions for vertices.
}
