% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{plot.qcc}
\alias{plot.qcc}
\title{Function to plot Shewhart chart for GraphAD}
\usage{
\method{plot}{qcc}(x, l = l, title, plot.LCL = FALSE)
}
\arguments{
\item{x}{a qcc object}

\item{l}{length of previous graphs to estimate moving averages and moving standard deviation.}

\item{title}{a string of the plot title}

\item{plot.LCL}{a Boolean variable to decide whether to show the anomalies lower than lower limits (LCL \eqn{\mu^{t}-3\sigma^t}) .}
}
\value{
A control chart ggplot for GraphAD
}
\description{
Function to plot Shewhart chart for GraphAD
}
