% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{plot.qcc.vertex}
\alias{plot.qcc.vertex}
\title{Function to plot Shewhart chart for VertexAD}
\usage{
\method{plot}{qcc.vertex}(x, l = l, title, plot.LCL = FALSE)
}
\arguments{
\item{x}{a list of qcc object}

\item{l}{length of previous graphs to estimate moving averages and moving standard deviation.}

\item{title}{a string of the plot title}

\item{plot.LCL}{a Boolean variable to decide whether to show the anomalies lower than lower limits (LCL \eqn{\mu_i^{t}-3\sigma_i^t}) .}
}
\value{
A control chart ggplot for VertexAD
}
\description{
Function to plot Shewhart chart for VertexAD
}
