% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pltclique}
\alias{pltclique}
\title{create a planted clique
It creates planted clique for a specific graph for a list of graphs.}
\usage{
pltclique(gip, p, art.anomaly.v)
}
\arguments{
\item{gip}{a list of graphs in \code{igraph} format.}

\item{p}{is the index of graph to be inserted a planted clique}

\item{art.anomaly.v}{is the vertex index in igraph.vs format to be planted clique.}
}
\value{
A list containing a planted clique size as size of art.anomaly.v at p-th graph
}
\description{
create a planted clique
It creates planted clique for a specific graph for a list of graphs.
}
\author{
Guodong Chen <gchen35@jhu.edu>
}
