% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{principal_components}
\alias{principal_components}
\title{Principal Component Analysis}
\usage{
principal_components(data)
}
\arguments{
\item{data}{numeric data}
}
\value{
A list containing:
    \enumerate{
      \item \code{eigval}: numeric vector of computed eigenvalues
      \item \code{eigvec}: numerical matrix of computed eigenvectors
      \item \code{pca_loadings}: numerical matrix of computed loadings
      \item \code{pca_scores}: numerical matrix of computed component (aka factor) scores
    }
}
\description{
\code{principal_components} relates the data to a set of a components through
the eigen-decomposition of the correlation matrix, where each component explains
some variance of the data. The eigenvalues can be normalized to determine how
much of the variance each component represents.
}
\examples{
x <- matrix(rnorm(200 * 3), ncol = 10)
principal_components(x)

}
