% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{principal_components_result}
\alias{principal_components_result}
\title{Easy Access to Principal Component Analysis Results}
\usage{
principal_components_result(data, results = 1)
}
\arguments{
\item{data}{list output from \code{principal_components}}

\item{results}{principal component analysis results to extract. Can use either
results name or number (i.e. eigvec or 2):
  \enumerate{
    \item \code{eigval} (default)
    \item \code{eigvec}
    \item \code{pca_loadings}
    \item \code{pca_scores}
  }}
}
\value{
Returns one of the selected results:
    \enumerate{
      \item \code{eigval}: numeric vector of computed eigenvalues
      \item \code{eigvec}: numerical matrix of computed eigenvectors
      \item \code{pca_loadings}: numerical matrix of computed loadings
      \item \code{pca_scores}: numerical matrix of computed component (aka factor) scores
    }
}
\description{
\code{principal_components_result} Provides easy access to principal
component analysis results
}
\examples{

# An efficient means for getting principal component analysis results
x <- matrix(rnorm(200 * 3), ncol = 10)

principal_components(x) \%>\%
  principal_components_result(pca_scores)

}
\seealso{
\code{\link{principal_components}} for computing the principal components results
}
