% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{project_networks}
\alias{project_networks}
\title{Function to estimated the score matrices of a list of graphs given the common invariant subspace V}
\usage{
project_networks(Adj_list, V)
}
\arguments{
\item{Adj_list}{list of adjacency matrices, of size n x n}

\item{V}{common invariant subspace. A matrix of size n x d.}
}
\value{
A list containing the score matrices
}
\description{
Function to estimated the score matrices of a list of graphs given the common invariant subspace V
}
