% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{qcc}
\alias{qcc}
\title{Main function to create a 'qcc' object}
\usage{
qcc(
  data,
  type = c("xbar", "R", "S", "xbar.one", "p", "np", "c", "u", "g"),
  sizes,
  center,
  std.dev,
  limits,
  data.name,
  labels,
  newdata,
  newsizes,
  newdata.name,
  newlabels,
  nsigmas = 3,
  confidence.level,
  rules = shewhart.rules,
  plot = TRUE,
  plot.LCL = FALSE,
  ...
)
}
\arguments{
\item{plot.LCL}{a Boolean variable to decide whether to show the anomalies lower than lower limits (LCL \eqn{\mu^{t}-3\sigma^t}) .}
}
\description{
Main function to create a 'qcc' object
}
\references{
see R package qcc
}
