% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilAD.R
\name{qccAD}
\alias{qccAD}
\title{Function to perform anomaly detection for time series of graphs}
\usage{
qccAD(
  glist,
  method = "OMNI",
  diag.augment = TRUE,
  l = 3,
  d = NULL,
  dsvd = d,
  approx = TRUE,
  par = FALSE,
  numpar = 2,
  elbow = 2,
  plot.figure = TRUE,
  plot.LCL = FALSE
)
}
\arguments{
\item{glist}{a list of undirected simple graphs (simple graphs are graphs which do not contain self-loop and multiple edges)
in igraph format with same number of vertices with vertices are 1-1 matched.
Graphs can be weighted or binary. (Say the length of list to be tmax)}

\item{method}{a character variable  to be chosen among c("OMNI","MASE").
The code will first do OMNIbus embedding (OMNI) or Multiple Adjacency Spectrally Ebedding (MASE) with two adjcaency matrices(can be weighted or not) of all input adjacent graphs sequentially.
Then use latent positions to calculate test statistics \eqn{y^{t}=||X^{t}- X^{t+1}||} using operator norm.
Then for \eqn{t=l,...,tmax-1}, we calculate the moving means \eqn{\mu^t} and moving standard deviations \eqn{\sigma^t} at \eqn{t} by \eqn{y^{t-l+1},...,y^{t-1}}.
So only tmax-l time points are ploted as first l graphs have been used as estimating moving means and standard deviations.}

\item{diag.augment}{a Boolean variable to decide whether to do diagonal augmentation when performing adjacency spectral embedding. Default is TRUE.}

\item{l}{an integer of the number of graphs in time window in estimating the moving mean and moving standard deviation. l must be less than number of graphs and be greater than 3.}

\item{d}{a fixed integer of dimension to perform OMNI and individual ASE for MASE.  If d is NULL, then dimension is chosen automatically.}

\item{dsvd}{An integer number of dimension only used in joint embedding for MASE. If NULL, then dimension is chosen automatically as second elbow selected in Zhu & Ghodsi method for the scree plot of the singular values of the concatenated spectral embeddings of MASE.}

\item{approx}{a Boolean variable to decide whether to use irlba package to find a few approximate singular values and corresponding singular vectors of a matrix. Default is TRUE.}

\item{par}{a Boolean variable to decide whether to run in parallel. Default is FALSE.}

\item{numpar}{an integer number to decide number of clusters for parallel implmentation. Default is 2.}

\item{elbow}{number of elbow in Zhu & Ghodsi method for the scree plot of each individual graph singular values for MASE or of each omnibus matrix singular values for OMNI.}

\item{plot.figure}{a Boolean variable to decide whether plot control chart. Default is TRUE.}

\item{plot.LCL}{a Boolean variable to decide whether to show the anomalies lower than lower limits (LCL \eqn{\mu^{t}-3\sigma^t}) .}
}
\value{
A list containing a vector GraphAD of length tmax-l which consists of control charts deviations, with the
a list VertexAD (length tmax-l) with vectors of anomalous vertices indices for each graph.
}
\description{
Function to perform anomaly detection for time series of graphs
}
\examples{
glist <- list()
for (i in 1:5) {
  glist[[i]] <- sample_gnp(100,.1)
  }
  glist[[6]] <- sample_gnp(100,.9)
  glist[[7]] <- sample_gnp(100,.1)
  for (i in 8:12) {
    glist[[i]] <- sample_gnp(100,.1)
    }
  result<- qccAD(glist, l=4,d=1,dsvd=1,method="OMNI",
  diag.augment = TRUE,approx=FALSE, par=FALSE, numpar=2)
  print(result.OMNI$GraphAD) #print the number of deviation for GraphAD, only positive ones are meaningful

# Sample a time series of RDPG graph (length tmax > 17) with same 1-1 matched vertices unweighted
# hollow symmetric undirected graphs, the latent positions i.i.d uniform.
# Some vertices in 16-th and 17-th graphs are given perturbations so there exists anomalies at 16:17.
n <- 100 #number of vertices
nperturb <- 20 #number of perturbed vertices
cperturb <- .12 #number of perturbation, larger cperturb means more obvious anomalies.
rmin <- .2 # parameter for uniform[rmin, rmax].
rmax <- .8 # parameter for uniform[rmin, rmax].
tmax <- 22 # number of graphs must be greater than 17.
#Generate data or load the data you want
glist <- generate.tsg(n, nperturb, cperturb=NULL, rmin, rmax, tmax)$glist
#Do anomaly detection with OMNI in parallel
result.OMNI <- qccAD(glist, l=11,d=1,dsvd=NULL,method="OMNI",
                     diag.augment = TRUE, approx=FALSE, par=TRUE, numpar=2)
#print the number of deviation for GraphAD, only positive ones are meaningful
print(result.OMNI$GraphAD)

# Do anomaly detection with MASE in parallel
result.MASE<- qccAD(glist, l=11,d=1,dsvd=2,method="MASE",
                    diag.augment = TRUE, approx=FALSE, par=TRUE, numpar=2)
#print the number of deviation for GraphAD, only positive ones are meaningful
print(result.MASE$GraphAD)

}
\references{
Zhu, Mu and Ghodsi, Ali (2006), Automatic dimensionality selection from
the scree plot via the use of profile likelihood, Computational
Statistics & Data Analysis, Volume 51 Issue 2, pp 918-930, November, 2006.

Levin, K., Athreya, A., Tang, M., Lyzinski, V. and Priebe, C.E., 2017, November. A central limit theorem for an omnibus embedding of multiple random dot product graphs. In 2017 IEEE International Conference on Data Mining Workshops (ICDMW) (pp. 964-967). IEEE.

Arroyo, J., Athreya, A., Cape, J., Chen, G., Priebe, C.E. and Vogelstein, J.T., 2019. Inference for multiple heterogeneous networks with a common invariant subspace. arXiv preprint arXiv:1906.10026.
}
