% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_for_testing.R
\docType{data}
\name{raw_data_time_gaps}
\alias{raw_data_time_gaps}
\title{raw_data_time_gaps}
\usage{
data(raw_data_time_gaps)
}
\description{
A data frame containing a time series with headings timestamp and count.
}
