% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rdpg.sample}
\alias{rdpg.sample}
\title{Sample RDPG graph with latent position}
\usage{
rdpg.sample(X)
}
\arguments{
\item{X}{latent position matrix of \eqn{d} columns and \eqn{n} rows.}
}
\value{
A un-directed hollow symmeric unweighted graph generated from bernoulli \eqn{EA=P=XX^{T}}
}
\description{
Sample RDPG graph with latent position
}
