% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulated_state_vector.R
\name{tabulate_state_vector}
\alias{tabulate_state_vector}
\title{Tabulate State Vector}
\usage{
tabulate_state_vector(data, blocks, num_IPs, IP_names = NULL)
}
\arguments{
\item{data}{data}

\item{blocks}{integer value to divide data}

\item{num_IPs}{integer value for the number of IP addresses to investigate}

\item{IP_names}{list of names for IP columns}
}
\value{

}
\description{
\code{tabulate_state_vector} divides data into blocks, where each block is
defined by descriptive statistics. Taking a contingency table approach, this
function separates variables of type factor into their unique factor levels
and counts the number of occurrences for those levels within each block. Due
to the large number of unique IP address, this function allows for the user
to determine how many IP addresses they would like to investigate (takes the
top occurrences for IP variables).
}
\examples{


}
