% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalyDetection.R
\docType{package}
\name{anomalyDetection}
\alias{anomalyDetection}
\alias{anomalyDetection-package}
\title{anomalyDetection: An R package for implementing augmented network log anomoly
detection procedures.}
\description{
anomalyDetection: An R package for implementing augmented network log anomoly
detection procedures.
}
